requirements.txt
aiogram==3.20.0
tortoise-orm==0.25.1
colorlog==6.9.0
aiohttp==3.12.13
Pillow==11.2.1
python-dotenv==1.1.0
asyncpg==0.30.0